fixed schemas, fixed seeder, added truncate and seed scripts

+++ migrations/20200314202829-create-payroll.js
@@ -17,11 +17,11 @@ module.exports = {
         allowNull: true
       },
       hoursWorked: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.REAL,
         allowNull: true
       },
       totalWage: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.REAL,
         allowNull: true
       },
       userId: {
